feat & fix : edit login/register uri & imageFile communication
