+++ client/WorkOutWithClient/app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         tools:targetApi="31"
         android:usesCleartextTraffic="true"> <!--Allow Http Protocol By dong-->
         <activity
-            android:name=".retrofitinterface.MainActivity"
+            android:name=".userinterface.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
