--- client/WorkOutWithClient/app/src/main/java/com/bubu/workoutwithclient/retrofitinterface/MainActivity.kt
@@ -1,12 +1,0 @@
-package com.bubu.workoutwithclient.retrofitinterface
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.bubu.workoutwithclient.R
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
