+++ client/WorkOutWithClient/app/src/main/java/com/bubu/workoutwithclient/retrofitinterface/UserChangePasswordModule.kt
@@ -1,9 +1,7 @@
-package com.bubu.workoutwithclient.userinterface
+package com.bubu.workoutwithclient.retrofitinterface
 
 import android.util.Log
-import com.google.gson.Gson
 import com.google.gson.JsonObject
-import com.google.gson.annotations.SerializedName
 import retrofit2.Call
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -13,12 +11,12 @@ import retrofit2.http.POST
 import java.io.EOFException
 import java.net.SocketTimeoutException
 
+data class UserChangePasswordData(val password1 : String, val password2 : String)
 
-data class UserSendEmailData(val email : String)
+class UserChangePasswordModule(override val userData: UserChangePasswordData)
+    : UserApiInterface {
 
-class UserSendEmailModule(override val userData: UserSendEmailData) : UserApiInterface {
-
-    interface UserSendEmailInterface {
+    interface UserChangePasswordInterface {
         @Headers("Content-Type: application/json")
         @POST("/v1/user/")
         fun get(
@@ -26,17 +24,16 @@ class UserSendEmailModule(override val userData: UserSendEmailData) : UserApiInt
         ): Call<Any>
         //보내는 데이터 형식
     }
-
-
     override suspend fun getApiData(): Any? {
         val retrofit = Retrofit.Builder()
             .baseUrl(super.serverAddress)
             .addConverterFactory(GsonConverterFactory.create())
             .build()
-        val retrofitObject = retrofit.create(UserSendEmailInterface::class.java)
+        val retrofitObject = retrofit.create(UserChangePasswordInterface::class.java)
         try {
             val requestData = JsonObject()
-            requestData.addProperty("email",userData.email)
+            requestData.addProperty("password1",userData.password1)
+            requestData.addProperty("password2",userData.password2)
             var resp = retrofitObject.get(requestData).execute()
             if (resp.code() in 100..199) {
                 return super.handle100(resp)
