+++ client/WorkOutWithClient/app/src/main/java/com/bubu/workoutwithclient/retrofitinterface/UserEmailAuthModule.kt
@@ -1,4 +1,4 @@
-package com.bubu.workoutwithclient.userinterface
+package com.bubu.workoutwithclient.retrofitinterface
 
 import android.util.Log
 import com.google.gson.Gson
