+++ client/WorkOutWithClient/app/src/main/java/com/bubu/workoutwithclient/retrofitinterface/UserGetCommunityListModule.kt
@@ -1,40 +1,41 @@
-package com.bubu.workoutwithclient.userinterface
+package com.bubu.workoutwithclient.retrofitinterface
 
 import android.util.Log
 import com.google.gson.Gson
-import com.google.gson.JsonObject
 import com.google.gson.annotations.SerializedName
 import retrofit2.Call
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.*
+import retrofit2.http.GET
+import retrofit2.http.Query
 import java.io.EOFException
 import java.net.SocketTimeoutException
 
-data class UserGetMatchRoomResponseData(
-    @SerializedName("userInfo") val userInfo: List<UserGetMatchRoomUserInfo>,
-    @SerializedName("voteInfo") val voteInfo: List<UserGetMatchRoomVoteInfo>
+data class UserGetCommunityListResponseData(@SerializedName("postList") val postList: List<PostList>)
+
+data class PostList(
+    @SerializedName("id") val id: String,
+    @SerializedName("postId") val postId: Int,
+    @SerializedName("title") val title: String,
+    @SerializedName("picture") val picture: String,//file
+    @SerializedName("timestamp") val timestamp: String,
+    @SerializedName("likeCount") val likeCount: Int
 )
 
-data class UserGetMatchRoomUserInfo(
-    @SerializedName("profilePic") val profilePic: String,//file
-    @SerializedName("name") val name: String
-)
+//postlist : [
+//id : 게시글 작성자 id
+//postid : 게시글 id
+//title : 제목
+//picture : 게시글 사진
+//timestamp : 게시글 작성 시간
+//likecount : 좋아요 수
+//]
 
-data class UserGetMatchRoomVoteInfo(
-    @SerializedName("voteTitle") val voteTitle: String,
-    @SerializedName("voteId") val voteId: Int
-)
+class UserGetCommunityListModule(override val userData: Any?) : UserApiInterface {
 
-data class UserGetMatchRoomData(val matchId: Int)
-
-class UserGetMatchRoomModule(override val userData: UserGetMatchRoomData) : UserApiInterface {
-    interface UserGetMatchRoomInterface {
+    interface UserGetCommunityListInterface {
         //@Headers("Content-Type: application/json")
-        @GET("/v1/matching/info")
+        @GET("/v1/community/")
         fun get(
-            @Query("token") token: String,
-            @Query("matchId") matchId: Int
+            @Query("token") token : String
         ): Call<Any>
         //보내는 데이터 형식
     }
@@ -45,38 +46,19 @@ class UserGetMatchRoomModule(override val userData: UserGetMatchRoomData) : User
             val result = auth.getApiData()
             if (result == true) {
                 val retrofit = ApiClient.getApiClient()
-                val retrofitObject = retrofit.create(UserGetMatchRoomInterface::class.java)
+                val retrofitObject = retrofit.create(UserGetCommunityListInterface::class.java)
                 try {
-                    var resp =
-                        retrofitObject.get(userInformation.accessToken, userData.matchId).execute()
+                    var resp = retrofitObject.get(UserData.accessToken).execute()
                     if (resp.code() in 100..199) {
                         return super.handle100(resp)
                     } else if (resp.code() in 200..299) {
                         val responseBody = super.handle200(resp)
-                        val jsonString: String = Gson().toJsonTree(responseBody).asJsonObject.toString()
-                        return convertToClass(jsonString, UserGetMatchRoomResponseData::class.java)
-                        /*val userInfoList = mutableListOf<UserGetMatchRoomUserInfo>()
-                        val voteInfoList = mutableListOf<UserGetMatchRoomVoteInfo>()
-
-                        val userInfoJsonArray = jsonObject.asJsonObject.getAsJsonArray("userInfo")
-                        val voteInfoJsonArray = jsonObject.asJsonObject.getAsJsonArray("voteInfo")
-                        userInfoJsonArray.forEach {
-                            userInfoList.add(
-                                UserGetMatchRoomUserInfo(
-                                    it.asJsonObject.get("profilePic").toString(),
-                                    it.asJsonObject.get("name").toString()
-                                )
-                            )
-                        }
-                        voteInfoJsonArray.forEach {
-                            voteInfoList.add(
-                                UserGetMatchRoomVoteInfo(
-                                    it.asJsonObject.get("voteTitle").toString(),
-                                    it.asJsonObject.get("voteId").toString().toInt()
-                                )
-                            )
-                        }
-                        return UserGetMatchRoomResponseData(userInfoList, voteInfoList)*/
+                        val jsonString: String =
+                            Gson().toJsonTree(responseBody).asJsonObject.toString()
+                        return convertToClass(
+                            jsonString,
+                            UserGetCommunityListResponseData::class.java
+                        )
                     } else if (resp.code() in 300..399) {
                         return super.handle300(resp)
                     } else if (resp.code() in 400..499) {
