+++ client/WorkOutWithClient/app/src/main/java/com/bubu/workoutwithclient/retrofitinterface/UserIsLikeCommunityModule.kt
@@ -1,29 +1,24 @@
-package com.bubu.workoutwithclient.userinterface
+package com.bubu.workoutwithclient.retrofitinterface
 
 import android.util.Log
-import com.google.gson.JsonObject
+import com.google.gson.Gson
 import com.google.gson.annotations.SerializedName
 import retrofit2.Call
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.Body
-import retrofit2.http.Headers
-import retrofit2.http.POST
-import retrofit2.http.PUT
+import retrofit2.http.GET
+import retrofit2.http.Query
 import java.io.EOFException
 import java.net.SocketTimeoutException
 
+data class UserIsLikeCommunityResponseData(@SerializedName("likeable") val likeable: Boolean)
 
-data class UserStartRoomVoteData(val matchId: String, val voteId: String)
 
-class UserStartRoomVoteModule(override val userData: UserStartRoomVoteData) : UserApiInterface {
+class UserIsLikeCommunityModule(override val userData: UserLikeCommunityData) : UserApiInterface {
 
-    interface UserStartRoomVoteInterface {
-        @Headers("Content-Type: application/json")
-        @PUT("/v1/matching/vote/")
+    interface UserIsLikeCommunityInterface {
+        @GET("/v1/community/like")
         fun get(
-            //@Query("targetId") targetId : String
-            @Body body: JsonObject
+            @Query("token") token: String,
+            @Query("postId") postId: String
         ): Call<Any>
         //보내는 데이터 형식
     }
@@ -34,17 +29,20 @@ class UserStartRoomVoteModule(override val userData: UserStartRoomVoteData) : Us
             val result = auth.getApiData()
             if (result == true) {
                 val retrofit = ApiClient.getApiClient()
-                val retrofitObject = retrofit.create(UserStartRoomVoteInterface::class.java)
+                val retrofitObject = retrofit.create(UserIsLikeCommunityInterface::class.java)
                 try {
-                    val requestData = JsonObject()
-                    requestData.addProperty("token", userInformation.accessToken)
-                    requestData.addProperty("matchId", userData.matchId)
-                    requestData.addProperty("voteId", userData.voteId)
-                    var resp = retrofitObject.get(requestData).execute()
+                    var resp =
+                        retrofitObject.get(userInformation.accessToken, userData.postId).execute()
                     if (resp.code() in 100..199) {
                         return super.handle100(resp)
                     } else if (resp.code() in 200..299) {
-                        return resp.code()
+                        val responseBody = super.handle200(resp)
+                        val jsonString: String =
+                            Gson().toJsonTree(responseBody).asJsonObject.toString()
+                        return convertToClass(
+                            jsonString,
+                            UserIsLikeCommunityResponseData::class.java
+                        )
                     } else if (resp.code() in 300..399) {
                         return super.handle300(resp)
                     } else if (resp.code() in 400..499) {
