+++ client/WorkOutWithClient/app/src/main/java/com/bubu/workoutwithclient/retrofitinterface/UserLikeCommunityModule.kt
@@ -1,44 +1,22 @@
-package com.bubu.workoutwithclient.userinterface
+package com.bubu.workoutwithclient.retrofitinterface
 
 import android.util.Log
-import com.google.gson.Gson
 import com.google.gson.JsonObject
-import com.google.gson.annotations.SerializedName
 import retrofit2.Call
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.GET
-import retrofit2.http.Query
+import retrofit2.http.*
 import java.io.EOFException
 import java.net.SocketTimeoutException
 
-data class UserGetCommunityListResponseData(@SerializedName("postList") val postList: List<PostList>)
+data class UserLikeCommunityResponseData(val code: Int)
+data class UserLikeCommunityData(val postId: String)
 
-data class PostList(
-    @SerializedName("id") val id: String,
-    @SerializedName("postId") val postId: Int,
-    @SerializedName("title") val title: String,
-    @SerializedName("picture") val picture: String,//file
-    @SerializedName("timestamp") val timestamp: String,
-    @SerializedName("likeCount") val likeCount: Int
-)
-
-//postlist : [
-//id : 게시글 작성자 id
-//postid : 게시글 id
-//title : 제목
-//picture : 게시글 사진
-//timestamp : 게시글 작성 시간
-//likecount : 좋아요 수
-//]
-
-class UserGetCommunityListModule(override val userData: Any?) : UserApiInterface {
-
-    interface UserGetCommunityListInterface {
-        //@Headers("Content-Type: application/json")
-        @GET("/v1/community/")
+class UserLikeCommunityModule(override val userData: UserLikeCommunityData) : UserApiInterface {
+    interface UserLikeCommunityInterface {
+        @Headers("Content-Type: application/json")
+        @POST("/v1/community/like/")
         fun get(
-            @Query("token") token : String
+            //@Query("token") token : String
+            @Body body: JsonObject
         ): Call<Any>
         //보내는 데이터 형식
     }
@@ -49,19 +27,16 @@ class UserGetCommunityListModule(override val userData: Any?) : UserApiInterface
             val result = auth.getApiData()
             if (result == true) {
                 val retrofit = ApiClient.getApiClient()
-                val retrofitObject = retrofit.create(UserGetCommunityListInterface::class.java)
+                val retrofitObject = retrofit.create(UserLikeCommunityInterface::class.java)
                 try {
-                    var resp = retrofitObject.get(userInformation.accessToken).execute()
+                    val requestData = JsonObject()
+                    requestData.addProperty("token", userInformation.accessToken)
+                    requestData.addProperty("postId", userData.postId)
+                    var resp = retrofitObject.get(requestData).execute()
                     if (resp.code() in 100..199) {
                         return super.handle100(resp)
                     } else if (resp.code() in 200..299) {
-                        val responseBody = super.handle200(resp)
-                        val jsonString: String =
-                            Gson().toJsonTree(responseBody).asJsonObject.toString()
-                        return convertToClass(
-                            jsonString,
-                            UserGetCommunityListResponseData::class.java
-                        )
+                        return resp.code()
                     } else if (resp.code() in 300..399) {
                         return super.handle300(resp)
                     } else if (resp.code() in 400..499) {
@@ -115,4 +90,5 @@ class UserGetCommunityListModule(override val userData: Any?) : UserApiInterface
             return e
         }
     }
+
 }
