+++ client/WorkOutWithClient/app/src/main/java/com/bubu/workoutwithclient/retrofitinterface/UserRegisterModule.kt
@@ -1,9 +1,8 @@
-package com.bubu.workoutwithclient.userinterface
+package com.bubu.workoutwithclient.retrofitinterface
 
 import android.util.Log
 import com.google.gson.Gson
 import com.google.gson.JsonObject
-import com.google.gson.annotations.SerializedName
 import retrofit2.Call
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -39,7 +38,7 @@ class UserRegisterModule(override val userData: UserRegisterData) : UserApiInter
 
     interface UserRegisterInterface {
         @Headers("Content-Type: application/json")
-        @POST("/api/accounts/v1/registration/")
+        @POST("/v1/registration/")
         fun get(
             @Body body: JsonObject
         ): Call<Any>
@@ -69,11 +68,7 @@ class UserRegisterModule(override val userData: UserRegisterData) : UserApiInter
             if (resp.code() in 100..199) {
                 return super.handle100(resp)
             } else if (resp.code() in 200..299) { //Successful!!
-                var responseBody = super.handle200(resp)
-                val jsonObject: JsonObject = Gson().toJsonTree(responseBody).asJsonObject
-                val accessToken = jsonObject.get("access_token").toString()
-                val refreshToken = jsonObject.get("refresh_token").toString()
-                return UserLoginResponseData(accessToken, refreshToken)
+                return resp.code()
             } else if (resp.code() in 300..399) {
                 return super.handle300(resp)
             } else if (resp.code() in 400..499) {
