+++ client/WorkOutWithClient/app/src/main/java/com/bubu/workoutwithclient/retrofitinterface/UserStartMatchModule.kt
@@ -1,30 +1,28 @@
-package com.bubu.workoutwithclient.userinterface
+package com.bubu.workoutwithclient.retrofitinterface
 
 import android.util.Log
 import com.google.gson.Gson
-import com.google.gson.JsonArray
 import com.google.gson.JsonObject
 import com.google.gson.annotations.SerializedName
 import retrofit2.Call
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.GET
-import retrofit2.http.Headers
-import retrofit2.http.Query
+import retrofit2.http.*
 import java.io.EOFException
 import java.net.SocketTimeoutException
 
-data class UserGetMatchListResponseData(
-    @SerializedName("matchId") val id: String,
-    @SerializedName("title") val title: String
+data class UserStartMatchResponseData(@SerializedName("matchId") val matchId: Int)
+
+data class UserStartMatchData(
+    val city: String, val county: String, val district: String,
+    val game: Int
 )
 
-class UserGetMatchListModule(override val userData: Any?) : UserApiInterface {
+class UserStartMatchModule(override val userData: UserStartMatchData) : UserApiInterface {
 
-    interface UserGetMatchListInterface {
-        @GET("/v1/matching/")
+    interface UserStartMatchInterface {
+        @Headers("Content-Type: application/json")
+        @POST("/v1/matching/")
         fun get(
-            //@Query("token") token : String
+            @Body body: JsonObject
         ): Call<Any>
         //보내는 데이터 형식
     }
@@ -34,27 +32,22 @@ class UserGetMatchListModule(override val userData: Any?) : UserApiInterface {
             var auth = UserAuthModule(null)
             val result = auth.getApiData()
             if (result == true) {
-                //Do Any Operation or Jobs..
                 val retrofit = ApiClient.getApiClient()
-                val retrofitObject = retrofit.create(UserGetMatchListInterface::class.java)
+                val retrofitObject = retrofit.create(UserStartMatchInterface::class.java)
                 try {
-                    var resp = retrofitObject.get().execute()
-                    if (resp.code() in 100..199) { //
+                    val requestData = JsonObject()
+                    requestData.addProperty("token", UserData.accessToken)
+                    requestData.addProperty("city", userData.city)
+                    requestData.addProperty("county", userData.county)
+                    requestData.addProperty("district", userData.district)
+                    requestData.addProperty("game", userData.game)
+                    var resp = retrofitObject.get(requestData).execute()
+                    if (resp.code() in 100..199) {
                         return super.handle100(resp)
                     } else if (resp.code() in 200..299) {
                         val responseBody = super.handle200(resp)
-                        val jsonString : String = Gson().toJsonTree(responseBody).asJsonArray.toString()
-                        return Gson().fromJson(jsonString,Array<UserGetMatchListResponseData>::class.java).toList()
-                        /*val list = mutableListOf<UserGetMatchLists>()
-                        jsonObject.forEach {
-                            list.add(
-                                UserGetMatchLists(
-                                    it.asJsonObject.get("matchId").toString(),
-                                    it.asJsonObject.get("title").toString()
-                                )
-                            )
-                        }
-                        return UserGetMatchListResponseData(list)*/
+                        val jsonString: String = Gson().toJsonTree(responseBody).asJsonObject.toString()
+                        return convertToClass(jsonString, UserStartMatchResponseData::class.java)
                     } else if (resp.code() in 300..399) {
                         return super.handle300(resp)
                     } else if (resp.code() in 400..499) {
@@ -107,6 +100,6 @@ class UserGetMatchListModule(override val userData: Any?) : UserApiInterface {
             Log.d("Exception", e.toString())
             return e
         }
-    }
 
+    }
 }
